feat: MiniMemoCardを一時的なWidgetとして扱うように変更
